### 2025-03-18 #7918 @editorialbot
[PRE REVIEW]: `conmat`: generate synthetic contact matrices for a given age-stratified population
**Submitting author:** @njtierney (<a href="http://orcid.org/0000-0003-1460-8722">Nicholas John Tierney</a>)
**Repository:** https://github.com/idem-lab/conmat
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.2.9000
