+++ paper/paper.md
@@ -1,53 +1,42 @@
 ---
 title: '`conmat`: generate synthetic contact matrices for a given age-stratified population'
-authors:
-- affiliation: 1
-  name: Nicholas Tierney
-  orcid: 0000-0003-1460-8722
-- affiliation: 1,2
-  name: Nick Golding
-  orcid: 0000-0001-8916-5570
-- affiliation: 1,3
-  name: Aarathy Babu
-  orcid: 
-- affiliation: 4
-  name: Michael Lydeamore
-  orcid: 0000-0001-6515-827X
-- affiliation: 1,3
-  name: Chitra Saraswati
-  orcid: 0000-0002-8159-0414
-date: today
-bibliography: references.bib
-cite-method: biblatex
-tags:
-- epidemiology
-- R
-- infectious disease
-affiliations:
-- index: 1
-  name: Telethon Kids Institute
-- index: 2
-  name: Curtin University
-- index: 3
-  name: 
-- index: 4
-  name: Monash University
-execute: 
-  echo: true
-  cache: false
 format: 
-  pdf: 
-    keep-md: true
-    fig-height: 4
-    fig-align: center
-    fig-format: png
-    dpi: 300
   html: 
     keep-md: true
     fig-height: 4
     fig-align: center
     fig-format: png
     dpi: 300
+tags:
+  - epidemiology
+  - R
+  - infectious disease
+authors:
+  - name: Nicholas Tierney
+    orcid: 0000-0003-1460-8722
+    corresponding: true
+    affiliation: "1"
+  - name: Chitra Saraswati
+    orcid: 0000-0002-8159-0414
+    affiliation: "1"
+  - name: Aarathy Babu
+    orcid: 0000-0002-6982-5989
+    affiliation: "1"
+  - name: Michael Lydeamore
+    orcid: 0000-0001-6515-827X
+    affiliation: "3"
+  - name: Nick Golding
+    orcid: 0000-0001-8916-5570
+    affiliation: "1,2"
+affiliations:
+  - index: 1
+    name: "The Kids Research Institute Australia, WA, Australia"
+  - index: 2
+    name:  "University of Western Australia, WA, Australia"
+  - index: 3
+    name: "Department of Econometrics and Business Statistics, Monash University, VIC, Australia"
+date: today
+bibliography: paper.bib
 ---
 
 
@@ -58,6 +47,7 @@ format:
 ::: {.cell}
 
 :::
+
 
 
 # Summary
@@ -91,6 +81,7 @@ The `conmat` package was developed to fill the specific need of creating contact
 We will generate a contact matrix for Tasmania, a state in Australia, using a model fitted from the POLYMOD contact survey. We can get the age-stratified population data for Tasmania from the Australian Bureau of Statistics (ABS) with the helper function, `abs_age_state()`:
 
 
+
 ::: {.cell}
 
 ```{.r .cell-code}
@@ -99,8 +90,9 @@ head(tasmania)
 ```
 
 ::: {.cell-output .cell-output-stdout}
+
 ```
-# A tibble: 6 x 4 (conmat_population)
+# A tibble: 6 × 4 (conmat_population)
  - age: lower.age.limit
  - population: population
    year state lower.age.limit population
@@ -112,11 +104,15 @@ head(tasmania)
 5  2020 TAS                20      31641
 6  2020 TAS                25      34115
 ```
+
+
 :::
 :::
+
 
 
 We can then generate a synthetic contact matrix for Tasmania, by extrapolating the contact patterns between age groups learned from the POLYMOD study, using `extrapolate_polymod()`.
+
 
 
 ::: {.cell}
@@ -125,80 +121,12 @@ We can then generate a synthetic contact matrix for Tasmania, by extrapolating t
[…]
