### 2025-05-04 #7918 @csoneson
👋🏻 @njtierney - I will handle your submission and start by finding at least two suitable reviewers. If you have suggestions (see e.g the link in the first post above) feel free to let me know.

In parallel, I would like to ask you to try to shorten the paper a little - at the moment, it is quite a bit above the length indicated in the JOSS [guidelines](https://joss.readthedocs.io/en/latest/paper.html) (250-1,000 words). Thanks!
