### 2025-05-21 #7918 @csoneson
👋🏻 @giovannic, @brunaw, @LCB0B, @ccamara, @ankurankan - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> conmat: generate synthetic contact matrices for a given age-stratified population
#7918

You can find more information about our checklist-driven review process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

Thanks in advance!
