### 2025-06-01 #7918 @csoneson
@njtierney Thanks for the suggestion (just as a note, in the future please don't tag suggested reviewers directly, but allow me to get in contact with them).

While we are searching for reviewers, please also have a look at shortening the paper a bit (see [comment](https://github.com/openjournals/joss-reviews/issues/7918#issuecomment-2849281200) above) - you can generate a new proof with `@editorialbot generate pdf`.
