### 2025-06-02 #7918 @anthbapt
Hello,

I'm happy to review the paper.
Please let me know about the next step.

Best wishes,
Anthony

Le dim. 1 juin 2025 à 15:04, Charlotte Soneson ***@***.***> a
écrit :

> *csoneson* left a comment (openjournals/joss-reviews#7918)
> <https://github.com/openjournals/joss-reviews/issues/7918#issuecomment-2927271605>
>
> 👋🏻 @Bisaloo <https://github.com/Bisaloo>, @brunaw
> <https://github.com/brunaw>, @LCB0B <https://github.com/LCB0B>, @ccamara
> <https://github.com/ccamara>, @ankurankan <https://github.com/ankurankan>,
> @anthbapt <https://github.com/anthbapt>, @bsiepe
> <https://github.com/bsiepe> - would you be interested in reviewing this
> submission for the Journal of Open Source Software (JOSS)?
>
> conmat: generate synthetic contact matrices for a given age-stratified
> population
> #7918 <https://github.com/openjournals/joss-reviews/issues/7918>
>
> You can find more information about our checklist-driven review process
> here <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>
>
> Thanks in advance!
>
