### 2025-06-02 #7918 @Bisaloo
Hi @csoneson @njtierney :wave: :wave:; I am potentially interested in reviewing this package but won't be available until the end of the month.

On the plus side, I already vaguely know the package and I am very familiar with JOSS review process.

I could probably hand in my review during the first week of July. Please let me know if this would work for you. Otherwise, I'll have to pass on this one.
