### 2025-06-02 #8326 @editorialbot
[REVIEW]: `conmat`: generate synthetic contact matrices for a given age-stratified population
**Submitting author:** @njtierney (<a href="http://orcid.org/0000-0003-1460-8722">Nicholas John Tierney</a>)
**Repository:** https://github.com/idem-lab/conmat
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
