### 2025-06-02 #8326 @Bisaloo
@csoneson, could you please also set an editorialbot reminder for me? It looks like I can't do it myself as a reviewer.
