### 2025-06-05 #7918 @njtierney
> [@njtierney](https://github.com/njtierney) Thanks for the suggestion (just as a note, in the future please don't tag suggested reviewers directly, but allow me to get in contact with them).
>
> While we are searching for reviewers, please also have a look at shortening the paper a bit (see [comment](https://github.com/openjournals/joss-reviews/issues/7918#issuecomment-2849281200) above) - you can generate a new proof with `@editorialbot generate pdf`.

Hi @csoneson ! Thanks for this - my apologies for tagging the reviewer directly!

I will take a look at shortening the paper soon.

Also 👋 @Bisaloo !
