### 2025-06-27 #8326 @csoneson
👋🏻 @anthbapt, @njtierney - just wanted to check in to see how things are going here. @njtierney - have you had a chance to look into @anthbapt's question above, as well as the shortening of the paper? @anthbapt - let us know if something is blocking you at the moment, and when you have additional comments for the author. Thanks all!
