### 2025-07-04 fe7fcda
Add link to pkgdown website
To enable cross-linking from other pkgdown websites. See https://pkgdown.r-lib.org/articles/linking.html#across-packages
