### 2025-07-06 idem-lab/conmat#188 @Bisaloo
Consider switching readr dependency to minty
If I'm not mistaken, readr is used here as a dependency only for its `parse_number()` function.

A package that exports the same function with a much more limited scope (and therefore reduced dependencies) is minty: https://cran.r-project.org/package=minty

Since it's a drop-in reimplacement, I believe it may be worth making the switch.

Related to https://github.com/openjournals/joss-reviews/issues/8326.
