### 2025-07-07 1d91081
Add base url to pkgdown config
