+++ _pkgdown.yml
@@ -1,3 +1,4 @@
+url: https://idem-lab.github.io/conmat
 
 home:
   title: An R Package For Creating Synthetic Contact Matrices
