### 2025-07-07 idem-lab/conmat#189 @Bisaloo
Review the need for dontrun
https://github.com/search?q=repo%3Aidem-lab%2Fconmat%20dontrun&type=code

It is good for examples to run automatically, or to be easily runnable by users, when possible. The community recommendation currently is to avoid `\dontrun{}` unless absolutely necessary.

Here, I see some uses of `\dontrun{}` that could probably be swapped or eliminated:

- IIRC, long-running examples can be protected by `\donttest{}` instead of `\dontrun{}`
- errors can be protected by `try()` instead of `\dontrun{}`, as described in the [examples chapter of the R packages book](https://r-pkgs.org/man.html#sec-man-examples-errors)
- for plots, `if (interactive())` probably makes `\dontrun{}` unnecessary: https://github.com/idem-lab/conmat/blob/271b82c31a3e45a490384781bf82bdcebaacce85/R/autoplot.R#L13

Related to https://github.com/openjournals/joss-reviews/issues/8326
