### 2025-07-07 #8326 @csoneson
Thanks @Bisaloo for your review!

@anthbapt, @njtierney - did you have a chance to look at my comment above?

> 👋🏻 [@anthbapt](https://github.com/anthbapt), [@njtierney](https://github.com/njtierney) - just wanted to check in to see how things are going here. [@njtierney](https://github.com/njtierney) - have you had a chance to look into [@anthbapt](https://github.com/anthbapt)'s question above, as well as the shortening of the paper? [@anthbapt](https://github.com/anthbapt) - let us know if something is blocking you at the moment, and when you have additional comments for the author. Thanks all!
