### 2025-07-09 #8326 @njtierney
Thank you very much @csoneson and @anthbapt for your comments, and  @Bisaloo for your review!

I will be able to take a closer look at this from Tuesday, but am in the process of finalising some other work at the moment, my apologies - I really value this work because it is making the package better, more useful, and more robust. Unfortunately I have a high workload at the moment that I need to manage.
