### 2025-08-06 05eed96
first attempt at using parallel from purrr rather than furrr
