### 2025-08-08 246676c
update ggplot snapshot to include change from Num -> Log
