### 2025-08-08 e3f4816
use minty over readr to use `parse_number()`. Resolves #188
