### 2025-09-02 idem-lab/conmat#194 @njtierney
Add contributing.md file
Re: https://github.com/openjournals/joss-reviews/issues/8326#issuecomment-3184281468
