### 2025-09-02 njtierney/conmat#195 @njtierney
Shorten JOSS paper
As discussed in https://github.com/openjournals/joss-reviews/issues/8326
