### 2025-09-02 dfd2380
properly apply \donttest, try(), and if (interactive()){}
