### 2025-09-02 be1d116
add CONTRIBUTING.md - resolves #194
