### 2025-09-22 #8326 @njtierney
Hi @csoneson ! Just touching base. We are just awaiting some comments from a co-author regarding fixing a somewhat tricky bug that arose due to change in the mgcv package. Hopefully we should have this fixed by the end of this week or next.
