### 2025-10-25 d715f82
update to use `ti()` in interaction - should resolve #192
