### 2025-11-13 07df104
Shorten paper
