+++ paper/paper.qmd
@@ -131,17 +131,7 @@ The two optional covariates are included depending on which setting the model is
 
 Each cell in the resulting contact matrix (after back-transformation of the link function), indexed ($i$, $j$), is the predicted number of people in age group $j$ that a single individual in age group $i$ will have contact with per day. The sum over all age groups $j$ for a particular age group $i$ is the predicted total number of contacts per day for each individual of age group $i$.
 
-The six covariates are:
-
-- $|i-j|$, 
-- $|i-j|^{2}$, 
-- $i + j$, 
-- $i \times j$, 
-- $\text{max}(i, j)$ and 
-- $\text{min}(i, j)$.
-
-These covariates capture typical features of inter-person contact, where individuals primarily interact with people of similar age (the diagonals of the matrix), and with grandparents and/or children (the so-called 'wings' of the matrix). The key features of the relationship between the age groups, represented by spline transformations of the six covariates, are displayed in Figure \ref{fig-show-partial-plots} for the home setting. The spline-transformed $|i-j|$ and $|i-j|^{2}$ terms give the strong diagonal lines modelling people generally living with those of similar age and the intergenerational effect of parents and (faintly) grandparents with children. The spline-transformed $\text{max}(i, j)$ and $\text{min}(i, j)$ terms give the higher rates of at-home contact among younger people of similar ages and with their parents.
-
+The covariate capture typical features of inter-person contact, where individuals primarily interact with people of similar age (the diagonals of the matrix), and with grandparents and/or children (the so-called 'wings' of the matrix). The key features of the relationship between the age groups, represented by spline transformations of the six covariates, are displayed in Figure \ref{fig-show-partial-plots} for the home setting. 
 ```{r}
 #| label: partial-plots-create
 #| echo: false
@@ -217,14 +207,6 @@ The interface to the model formula in `conmat` is fixed; users cannot change the
 Public health decisions are often based on age specific information, which means the more accurate your age specific models are, the better those decisions are likely to be. This is the first piece of software that will provide appropriate contact matrices for a population, which means more accurate models of disease. 
 
 This code underlying this software was used as a key input into several models for COVID-19 transmission and control in Australia and contributed to decisions around vaccination policy [@DohertyModelling; @conway2023; @ryan2024].
-
-Some future directions for this software include:
-
-* integrating model fitting methods with contact survey prepared using the `socialmixr` R package [@socialmixr];
-* enabling prediction to arbitrary age brackets, e.g. monthly age bins for infants;
-* fitting to multiple contact surveys simultaneously, e.g., POLYMOD and CoMix;
-* providing estimates of uncertainty in contact matrices;
-* adding methods to include household size distributions in predictions of contacts in the 'home' setting;
 
 Software is never finished, and the software in its current format has proven useful for infectious disease modelling. In time we hope it can become more widely used and be useful for more applications in epidemiology and public health. 
 
