### 2025-11-28 17a436c
misc changes to improve documentation
