+++ vignettes/sir-model.Rmd
@@ -6,7 +6,8 @@ vignette: >
     %\VignetteEngine{knitr::rmarkdown}
     %\VignetteEncoding{UTF-8}
 ---
-```{r, include = FALSE}
+```{r include = FALSE}
+#| label: setup
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
