### 2025-11-28 7f83a8d
fix duplicate library label code chunk
