### 2025-11-28 45a038d
fix bug in future::plan where socket isn't closed - set sequential plan at the end of plan
