+++ man/raw_eigenvalue.Rd
@@ -20,9 +20,11 @@ Get raw eigvenvalue from NGM matrix
 \donttest{
 perth <- abs_age_lga("Perth (C)")
 perth_contact <- extrapolate_polymod(perth)
+perth_contact
+# must be the same age breaks as in perth_contact - 0-75
 perth_ngm <- generate_ngm(
   perth_contact,
-  age_breaks = c(seq(0, 85, by = 5), Inf)
+  age_breaks = c(seq(0, 75, by = 5), Inf)
 )
 raw_eigenvalue(perth_ngm)
 }
