### 2025-11-28 8f8e9fd
add age_breaks to transmission_probability_matrix()
