### 2025-11-28 d0bf46e
various qmd fixes
