+++ paper/paper.html.md
@@ -1,0 +1,198 @@
+---
+title: '`conmat`: generate synthetic contact matrices for a given age-stratified population'
+format: 
+  html: 
+    keep-md: true
+    fig-height: 4
+    fig-align: center
+    fig-format: png
+    dpi: 300
+tags:
+  - epidemiology
+  - R
+  - infectious disease
+authors:
+  - name: Nicholas Tierney
+    orcid: 0000-0003-1460-8722
+    corresponding: true
+    affiliation: "1"
+  - name: Chitra Saraswati
+    orcid: 0000-0002-8159-0414
+    affiliation: "1"
+  - name: Aarathy Babu
+    orcid: 0000-0002-6982-5989
+    affiliation: "1"
+  - name: Michael Lydeamore
+    orcid: 0000-0001-6515-827X
+    affiliation: "3"
+  - name: Nick Golding
+    orcid: 0000-0001-8916-5570
+    affiliation: "1,2"
+affiliations:
+  - index: 1
+    name: "The Kids Research Institute Australia, WA, Australia"
+  - index: 2
+    name:  "University of Western Australia, WA, Australia"
+  - index: 3
+    name: "Department of Econometrics and Business Statistics, Monash University, VIC, Australia"
+date: today
+bibliography: paper.bib
+---
+
+
+::: {.cell}
+
+:::
+
+
+
+::: {.cell}
+
+:::
+
+
+# Summary
+
+Contact matrices describe the number of contacts between individuals. They are used to create models of infectious disease spread. `conmat` is an R package which generates synthetic contact matrices for arbitrary input demography, ready for use in infectious disease modelling.
+
+There are currently few options for a user to access synthetic contact matrices [@socialmixr; @prem2017]. Existing code to generate synthetic contact matrices from @prem2017 are not designed for replicability, are restricted to select countries, and provide no sub-national demographic estimates.
+
+The `conmat` package exposes model fitting and prediction separately to the user. Users can fit a model based on a contact survey such as POLYMOD [@mossong2008], then predict from this model to their own demographic data. This means users can generate synthetic contact matrices for any region, with any contact survey.
+
+We demonstrate a use-case for `conmat` by creating contact matrices for sub-national level (in this case, a state) in Australia. 
+
+For users who do not wish to run the entire `conmat` pipeline, we have pre-generated synthetic contact matrices for 200 countries, based on a list of countries from the United Nations, using a model fit to the POLYMOD contact survey. These resulting synthetic contact matrices, and the associated code, can be found in the syncomat analysis pipeline ([GitHub](https://github.com/idem-lab/syncomat), [Zenodo](https://zenodo.org/records/11365943)) [@syncomat].
+
+# Statement of need
+
+Infectious diseases like influenza and COVID-19 spread via social contact. If we can understand patterns of contact---which individuals are more likely be in contact with each other---then we will be able to create models of how disease spreads. Epidemiologists and public policy makers can use these models to make decisions to keep a population safe and healthy.
+
+Empirical estimates of social contact are provided by social contact surveys. These provide samples of the frequency and type of social contact across different settings (home, work, school, other). 
+
+A prominent contact survey is the POLYMOD study by @mossong2008, which surveyed 8 European countries: Belgium, Germany, Finland, Great Britain, Italy, Luxembourg, The Netherlands, and Poland [@mossong2008]. 
+
+These social contact surveys can be projected on to a given demographic structure to produce estimated daily contact rates between age groups. These are known as 'synthetic' contact matrices. A widely used approach by @prem2017 [@prem2021] produced synthetic contact matrices for 177 countries at 'urban' and 'rural' levels for each country. 
+
[…]
