### 2025-11-29 #8326 @njtierney
Thanks @csoneson - what is the recommended word limit for JOSS? Are we required to get it under 1000 words as described in https://joss.readthedocs.io/en/latest/paper.html or do you have a target word limit we should aim for?
