### 2025-11-29 #8326 @csoneson
@njtierney The recommended word limit is 1,000 as stated in the guidelines - there is a bit of flexibility, but I think it would be great to get down to, say ~1,300 words or so, so a bit closer to the guidelines.
