### 2025-12-18 #8326 @csoneson
@njtierney - before I check further, please have a go at shortening the paper a bit as we have previously discussed. Thanks!
