### 2025-12-18 #8326 @csoneson
It seems it's not picking up the changes - is the updated paper pushed to the default branch on GitHub?
