### 2025-12-19 #8326 @njtierney
Hmmm! That is strange - here's the commit log: https://github.com/idem-lab/conmat/commits/master/ -
