### 2026-01-21 #8326 @MikeLydeamore
@csoneson The zenodo release should be available here: https://zenodo.org/records/18307331
Please let me know if there are issues or you need anything else. Apologies for the delay.
