### 2026-01-22 #8326 @csoneson
Thanks @MikeLydeamore! Most things look good, but the version tag does not seem to match between the release in the GitHub repository (`v1.0`) and the Zenodo record (`v1`). Could you look into this?
