### 2026-01-22 #8326 @MikeLydeamore
Sorry @csoneson I've updated the Zenodo record to v1.0.
