### 2026-01-23 #8326 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
