### 2026-01-30 #8326 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
As you can see, most seems in order, however the below are some points that require your attention (mostly small copyedits!) :point_down: :
- [x] Please remove the `text styling` from the paper title
- [ ] Third paragraph of Summary: a few corrections to this sentence, it should read "We demonstrate a use case for `conmat` by creating contact matrices for a sub-national entity (in this case, a state) in Australia."
- [x] Second paragraph of statement of need: "what is known" -> "what are known"
- [x] end of same paragraph: spell out the word "eight" and decapitalize "countries"
- [x] last paragraph of Statement of Need: hyphenate "next-generation"
- [x] first paragraph of Example: remove comma after "helper function"
- [x] last paragraph before conclusions: need a period at the end of last sentence (can add a space if you like to clarify that period is not part of URL)
- [x] second paragraph of conclusions: hyphenate "country-level"
- [x] last sentence before acknowledgements: add comma after "In time"
- [x] acknowledgements: elsewhere in the text, you have rendered `conmat` in monospace, suggest being consistent here
