### 2026-02-05 #8326 @MikeLydeamore
Hi @rkurchin thanks for pointing these out, I have made the requested changes to the paper in https://github.com/idem-lab/conmat/commit/1262f8c20d46ebc996ad2dbd2e72d0a7c07444ac
