### 2026-02-05 #8326 @rkurchin
🎉 Congrats to @njtierney and thank you to editor @csoneson and reviewers @anthbapt and @Bisaloo – JOSS couldn't run without your efforts!
